break down each digest into its own file
